Create Decoder for producing opcodes
