+++ src/gameboy/opcodes/run_fns/cp.rs
@@ -1,7 +1,8 @@
 use crate::gameboy::cpu::CPU;
 
 use super::super::super::{write_flag, Flags};
-use super::super::{Argument, RegisterLabel8};
+use super::super::Argument;
+use crate::gameboy::RegisterLabel8;
 
 pub fn run_cp(args: &[Argument], cpu: &mut CPU, memory: &mut Vec<u8>) -> u32 {
     // Clear all the flags
