+++ src/gameboy/tests/jump_test.rs
@@ -65,8 +65,8 @@ mod jump_test {
         gb.set_register_16(RegisterLabel16::ProgramCounter, 0x01);
 
         let cycles = gb.step_once();
-        use super::super::super::opcodes::decode_instruction;
-        let opcode = decode_instruction(0x01, gb.get_memory_slice_at(0, 0xFFFF)).unwrap();
+        use super::super::super::opcodes::Decoder;
+        let opcode = Decoder::decode_instruction(0x01, gb.get_memory_slice_at(0, 0xFFFF)).unwrap();
         assert_eq!(opcode.size(), 2);
 
         assert_eq!(cycles, 12);
